feat: 画像プレビューの遷移もRecyclerViewのコンポーネント内部で行わないようにした
