+++ modules/features/note/src/main/java/net/pantasystem/milktea/note/media/MediaPreviewHelper.kt
@@ -2,28 +2,22 @@
 package net.pantasystem.milktea.note.media
 
 import android.annotation.SuppressLint
-import android.app.Activity
 import android.view.LayoutInflater
 import android.view.View
 import android.widget.FrameLayout
 import android.widget.ImageButton
 import android.widget.ImageView
 import android.widget.TextView
-import androidx.core.app.ActivityOptionsCompat
 import androidx.core.view.isVisible
 import androidx.databinding.BindingAdapter
 import com.bumptech.glide.Glide
 import com.bumptech.glide.load.resource.bitmap.CenterCrop
-import dagger.hilt.android.EntryPointAccessors
-import dagger.hilt.android.internal.managers.FragmentComponentManager
 import jp.wasabeef.glide.transformations.BlurTransformation
 import net.pantasystem.milktea.common.glide.GlideApp
 import net.pantasystem.milktea.common.glide.blurhash.BlurHashSource
 import net.pantasystem.milktea.common_android.platform.isWifiConnected
 import net.pantasystem.milktea.common_android.ui.MediaLayout
 import net.pantasystem.milktea.common_android.ui.VisibilityHelper.setMemoVisibility
-import net.pantasystem.milktea.common_android_ui.NavigationEntryPointForBinding
-import net.pantasystem.milktea.common_navigation.MediaNavigationArgs
 import net.pantasystem.milktea.model.setting.Config
 import net.pantasystem.milktea.model.setting.MediaDisplayMode
 import net.pantasystem.milktea.note.R
@@ -48,33 +42,14 @@ object MediaPreviewHelper {
             return
         }
         val listener = View.OnClickListener {
-            val activity = FragmentComponentManager.findActivity(it.context) as Activity
-            val intent = EntryPointAccessors.fromActivity(
-                activity,
-                NavigationEntryPointForBinding::class.java
+            noteCardActionListenerAdapter?.onMediaPreviewClicked(
+                previewAbleFile = previewAbleFile,
+                files = previewAbleFileList,
+                index = previewAbleFileList.indexOfFirst { f ->
+                    f === previewAbleFile
+                },
+                thumbnailView = thumbnailView,
             )
-                .mediaNavigation().newIntent(
-                    MediaNavigationArgs.Files(
-                        files = previewAbleFileList.map { fvd ->
-                            fvd.source
-                        },
-                        index = previewAbleFileList.indexOfFirst { f ->
-                            f === previewAbleFile
-                        })
-                )
-            val context = it.context
-
-            if (context is Activity) {
-                val compat = ActivityOptionsCompat.makeSceneTransitionAnimation(
-                    context,
-                    thumbnailView,
-                    "image"
-                )
-                context.startActivity(intent, compat.toBundle())
-
-            } else {
-                context.startActivity(intent)
-            }
         }
         thumbnailView.setOnClickListener(listener)
         playButton.setOnClickListener(listener)
